refactor: Players에서 판정 로직 분리 후 Referee에 위임
